Porting magic zero to TensorIndexer (#4153)
This PR just ports the magic zero WAR to `TensorIndexer`. We should
revisit if magic zero is still meaningful, but for now, this is yet
another PR to fill remaining missing pieces of `TensorIndexer`.

I also ended up changing some of the functions I just added in #4158.
For example, `TensorIndexer::getUsedLoops`, which was added in #4158, is
replaced with `getUsedForLoopsOf` since `getUsedLoops` does not
guarantee a loop index is indeed used in the final index (e.g., when
unswitched, loop indices are replaced with 0 or `N - 1`, where `N` is
the extent of the iter domain).

NOTE: Magic zero has just been ignored in `TensorIndexer`. If it should
not be used, please explicitly disable it, which is already done in [the
matmul
scheduler](https://github.com/NVIDIA/Fuser/blob/main/csrc/scheduler/matmul_utils.cpp#L1035).