Translate MmaOp patterns properly on Hopper (#4072)
#3986 fixes our most common use cases of MatmulOp and LinearOp
translation on Hopper. It does so by scheduling global intermediates'
mtypes and allocation domains during _translation_. However, in case
there is no translation and we are already given an MmaOp this fails.
The current PR instead does mtype and allocation domain propagation
while caching operands, so that we can properly set `as_` and `bs_` and
so forth. This means that the input fusions don't need to differ between
hopper and ampere anymore, so we can translate both cases in the same
way and the only difference will be during scheduling.

Note that this will also make it easier to maintain internal tooling
which uses things like `canonicalizeInputToBMNK`.

Changes in this PR:
- Remove `avoid_intermediates` argument to
`MatmulPattern::translateToMmaOp` and update all call sites.
- Remove some helper utilities in `mma_utils.cpp`
- Introduce `scheduler_utils::scheduleInputToSkipIntermediates` which
will schedule the allocation domains and memory types of consumers of
inputs recursively to avoid "metadata ops" at the beginning of a fusion.
- Rearrange `HopperMultipleMatmulScheduler` to remove
`defineOperandCaches` and move `cacheInputsAndOutputs` after pattern
translation but before `findRoles`. Also `cacheInputsAndOutputs` now
uses `scheduler_utils::scheduleInputToSkipIntermediates` and defines the
operand roles as the last gmem tensor returned by that utility.
- Unguards `AllocationDomainTest.BasicMatmul/*` to allow it to run on
Hopper

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>
Co-authored-by: Christian Sarofeen <csarofeen@nvidia.com>
Co-authored-by: Jingyue Wu <wujingyue@gmail.com>
Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>