Insert block sync if it doesn't exist in WarAsyncWaitInserter (#4151)
I believe this check was to ensure that if a sync did exist in the
scope, that it was last, but no RAW sync is inserted when `blockDim.y ==
1`.

If there is no actual RAW sync, do we need the block sync to protect
against WAR for wgmma async? If so then we could instead check for its
existence in the scope and insert if needed.

Fixes #4138