Support computeWith in TensorIndexer (#4163)
This PR adds support of `computeWith` to IdModel. It's only used in [the
outer normalization
scheduler](https://github.com/NVIDIA/Fuser/blob/main/csrc/scheduler/normalization_utils.cpp#L1557-L1564),
and it may be considered outdated since its primary purpose is to
promote better compute and memory access overlapping, for which TMA and
warp specialization should be used on modern architectures. In the
meantime, the necessary change is minimal, so this PR is just for
keeping the compatibility with the legacy indexer.

Closes #4176