Expand ComputeAtMap exact and almost_exact disjointsets by IdModel (#4178)
`ComputeAtMap` is considered almost deprecated but the gap with
`IdModel` results in small but many changes in generated index exprs,
mainly because some mappings are missed in `ComputeAtMap`, which then
results in different index exprs.

To make the transition from the legacy indexer smoother, this PR
augments the exact and almost-exact disjoint sets of `ComputeAtMap` with
the exact and almost-exact graphs of `IdModel`, respectively. This
should make the code diff results cleaner when comparing the legacy and
the new indexers.