Move RAW and WAR sync insertion passes after circular buffering pass to support warp specialization. (#4187)
This PR moves the `insertRawThreadSynchronization` and
`insertWarThreadSynchronization` passes after `CircularBufferPass`.

## Why?
Warp-Specialized circular buffering creates `LoadWarp` and `ComputeWarp`
in an `IfThenElse` node. Adding a full block sync inside the
`IfThenElse` will cause the CUDA kernel to deadlock.

The `kir::BlockSync` node should be aware if it is in a warp-specialized
section of the code, so we can sync the appropriate amount of threads.

It is easier to identify this pattern if we insert block syncs after
creating the warp specialized `IfThenElse`.

```cuda
if (blockIdx.y == 2) {
  // Load Warp --- Only 128 threads
} else {
  // Compute Warp --- Remaining computer threads
}
```