Add missing device meshes in tests (#4117)
This PR is to decouple upcoming changes to the multi-device related
presegmentation passes.

Since the presegmentation passes will be updated serially, this to
workaround around test failures in the interim.
For example, some examples have manually set loop domains as required
for communication. The `propagateShardings` preseg pass will update the
loop domain if a tensorview does not have a device mesh. Since our
`reorderShardedAxis` preseg pass is not yet fully functional for DID
loop split, we can run into errors due to incomplete changes to the
preseg pass.

This PR is to ease rolling out the changes incrementally.

Eventually, we should not require setting meshes once the update to all
passes are complete.

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>