Make `getSwizzleMode` an `ir_utils` (#4185)
Some refactoring:
- Move `index.cpp::getSwizzleMode` to `ir_utils`, modify it to support
both TMA smem tv and MMA input tv.
- Replace `AllocationInfoMap::getTmaSwizzle` with
`ir_utils::isTMAOrMMASmemTv` + `ir_utils::getSwizzleMode`

For most cases, this change is a no-op. For the edge case where we are
using plain pointer load (instead of TMA) for loading the operand of
MMA, we are now ensuring the same alignment requirement. On Hopper, not
handling this edge case correctly does not trigger any error by luck.
But on Blackwell, we do not have such luck because we have an additional
smem tensor which is the tensor storing the TMem address for
`tcgen05::alloc`.