Do not map splits when input extent is 1 in AlmostExact (#4214)
This PR addresses a part of the issue #4218. I'll have another follow-up
PR.

This PR restricts mappings of the AlmostExact graph. Specifically, if a
split has an input ID with size one extent and its factor is not 1, the
input and the size-one output are currently mapped in AlmostExact. Just
found that this can result in invalid indexing. For example,

```
  // tv1: [r0(8)]
  tv1->split(0, 10);
  // [r1(1), r2(10)]
  tv1->split(0, 4);
  // [r3(1), r4(4), r2(10)]
```

When predicating this `tv1`, we want to generate an index as `i1 * 10 +
i2`, where `i1` and `i2` correspond to the loop indices of `r4` and
`r2`, respectively. However, with the current AlmostExact graph, this
would just be `i2` since `r1` and `r3` are mapped and the index or `r3`
is just 0 since its extent is 1.

Fundamentally, I think what we are missing is for patterns like this, it
is not valid to just pass the index of `r3` to `r1`, which effectively
means `r3` and `r1` should not be mapped.

With this PR, such split input and output IDs are no longer considered
trivially mappable. I'm not 100% confident that there's no unintended
side effect, but this seems to make most sense to me.