Use set allocation domains if used by TMA or MmaOp (#4234)
TensorIndexer generates incorrect indices for shared memory tensors that
are used by TMA. In that case, the set allocation domain should be
honored.

Similarly, set allocation domains are used when consumed by MmaOp. This
doesn't seem to be necessary for correctness, but it seems to make sense
to do so.