InsertReshardingsPass tries to rFactor DID reductions (#4209)
For #4133

This fixes InsertReshardingsPass for Allreduce. ReduceScatter needs
extra work that will be done in a separate PR.

This PR also lifts the current assumption that `r` can't be parallelized
on DID. There's nothing wrong with that assumption, and this PR merely
makes sharding propagation and decomposition more convenient. For
example, a matmul+allreduce can be decomposed by rFactor'ing the output
TensorView without even looking at the input TensorViews.