DID loop split for allgather for non-outermost sharded axis. (#4170)
Adds support for allgather if the sharded axis is not outermost.
`ProcessGroupNCCL` and `UCC` does require allocation of the sharded axis
to be outermost. We do not change the logical shape, and instead permute
the tensors to meet the requirements of NCCL and UCC within
`postAllgather`.

This will be added within the `reorderShardedAxis` preseg pass to
correctly set the loop and allocation domain for Allgather
communication. Additionally, a `set` operation is needed to change the
allocation of input if it does not have the sharded axis as the
outermost allocated axis.