[Host irs] alias and preallocated output support (#4144)
This PR belongs to a series of stacked PRs:
1. **=> You are here: #4144**
2. #4145
3. #4146
4. #4147

# What

- Support for aliases in HostIrContainer. When a Tensor tv1 is marked as
being the alias of tv0, then, at runtime, tv0's concrete data/buffer
will be used for the op. It is a way to reuse buffers that have been
allocated elsewhere within the TensorView's SSA paradigm. Chained
aliasing (tv2-->tv1-->tv0) are supported.
- Fix preallocated outputs in HostIrEvaluator 

# Why

It is necessary for stream parallelization, where typically we allocate
the full output buffer but each stream writes to a slice of this buffer.

# How 

The aliasing is stored in the HostIrContainer through a map.

At the HostIrEvaluator level, instead of operating directly on the
ExprEvaluator to write/read concrete data, we first apply the alias
indirection