[Host Ir] refactor and cleanup lowering and segmentation (#4145)
This PR belongs to a series of stacked PRs:
1. #4144
2. **=> You are here:** #4145
3. #4146
4. #4147

# What

1. We replace the bool option `SegmentCandidateFinderOptions::
only_segment_resharding_exprs` by a pointer to a predicate function.
This allow the user of the segmented to (optionally) provide a custom
function to be used to decide whether two given groups should be merged.
This achieves better separation of responsibility: with this option, the
segmented is only responsible of applying the segmentation algorithm,
but does not embed the specific rule for merging group which depends on
the application. The specific rule in our context is decided by the Hir
lowering. Imo this refactoring should ideally go further and make the
segmented a more abstract class that would be used in both Host Ir and
FusionExecutorCache lowering but only changing the newly introduced
function pointer.
2. In HIR lowering, we clearly separate (in a distinct for-loop, but
this later will become a preseg pass) the pass that transforms
resharding exprs into a Communication

# Why 
that's a preliminary refactoring useful for more advanced Host Ir
Lowering, notably ParallelType::Stream lowering