Add mutex guard to protect data race on options. (#4287)
Credit goes to @csarofeen for finding the segfault cause by
https://github.com/NVIDIA/Fuser/blob/5368ed00b65042f8797401412ea1f2d8a6a4cf5e/csrc/id_model/utils.h#L163-L166

We have observed a reliable segfault coming from
`MovePadTest.CascadePadCase2`.

This PR preserves the existing behavior, by adding a mutex to guard data
access on global options to protect us from racing.

An alternative is to make options thread_local, but we needed to figure
out how to inherit existing options on the forked threads. I added a
note on that.