check ID coverage for  reference_tv in reduction scheduler (#4223)
Fixes #3811  

Added a compile time check in reduction scheduler to ensure that all
output IDs are covered by the reference tv. This ensures that we will
not have dangling ID in the loop domain that's not scheduled.

Tasks:
- [x] verify that the added check doesn't break existing
tests/expectations on the capability of reduction scheduler: doesn't
seem to produce any code diff in CI.
- [x] what about other schedulers? Normalization scheduler should have
the same issue. But it's not reproducing the error, because the local TV
is fully inlined (compute at == -1).