Update propagateSharding preseg pass for DID loop split (#3838)
This PR extends the `propagateSharding` presegmentation pass for DID
loop splits.
Key changes:
1. We use TransformPropagator for all expressions except `ViewOp` which
is handled manually since TransformPropagator does not support it
without first propagating the reshape to the producer.
2. `makeReshardingContiguous` sets allocation domain for tvs with device
mesh. Ideally, we need to set it only for global tensors but this is not
known before segmentation, but should be set before segmentation.
3. ~The following tests are modified: See
[discussion](https://github.com/NVIDIA/Fuser/pull/3838#issuecomment-2807645305)~.
PR #4274 resolved this.

Follow-up PRs:

- `ViewOp` will be handled in a followup PR.
- Currently, we only backpropagate sharding for a tv that does not
already have a device dimension. This can be extended to propagate for
all parallel types not present on the tv. This will be done in a
followup. Backpropagating shardings can incorrectly change DIDx to
serial or modify DIDx to be on another location. `shardAllLike` can be
modified to specify which parallel type to propagate. Since
`insertResharding` and `propagateSharding` require different behavior, I
will handle it in a separate PR.
- Use `TransformReplay::CasP` in lieu of TransformPropagator.
- Propagate DID transforms within `castOp`:
[privatizeUpcast](https://github.com/NVIDIA/Fuser/blob/ed687366cf717837c8ea3e40f56542fec48e1616/csrc/fusion_segmenter.cpp#L4235-L4238)
clones cast operations, which fails segmentation since the transforms
are not replicated.

Findings from experiments:
https://github.com/NVIDIA/Fuser/pull/3838#issuecomment-2807645305

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>