Switch axis we use to compute swizzled_tiles (#4311)
#4242 turned on "grid traversal factor" which is a good thing. However,
it exposed a bug in how we limit that factor to prevent overrun in case
the swizzled axis has fewer tiles than the factor. This led to a
regression from 58% to 35% geomean perf compared to eager on H200.

This PR swaps the axes used to compute the number of swizzled tiles and
takes us from a geomean of 35% to 65% on
`benchmarks/python/test_matmul.py` on H200.