Switch rasterization order in non-persistent matmul kernels (#4316)
We currently parallelize like `[BIDx, BIDy, ...]` for
`mparams->cta_order == RowMajor`. This is incorrect as the dimensions of
the output should be interpreted as `[num_rows, num_cols]`, i.e.
row-major should have as its fast axis the _second_ dimension (which has
size `num_cols`). This PR swaps the meaning of row and column major for
both ampere and hopper schedulers.

Notice that cta order is specified intuitively in the default heuristic:

https://github.com/NVIDIA/Fuser/blob/fadfde5a44ed28998c99fcb9fd8e796c5b3de142/csrc/scheduler/matmul_utils.cpp#L400-L402.
This chooses as fast axis the shorted tile grid dimension, but needs
this PR to be effective.