Prefer merging pad with consumer groups (#4297)
~Not dependent but stacked on #4629~

This PR promotes merging pad exprs with consumers. Since padding expands
an iter domain, if possible, it would generaly result in more efficient
segmentation if the pad is grouped together with its following consumers
rather than its producers.

An actual motivating example is that, in the Litgpt Lllama RoPE
backward, there's a concatenation of three tensors, one of which ends up
having the pad expr in one segment, followed by a segment that consists
of the pad exprs of the other two tensors and the final cat expr. If all
of the pad exprs and the concat expr were grouped together, the
intermediate tensors between the two segments should have been much
smaller.