Fix `wgmma` sync insertion for data-parallel. (#4312)
## General
* Added a check for `warpgroup` warnings generated by `ptxas_verbose` to
`MLPBenchmarkTest`. This is very valuable for determining if the CUDA
kernel is correct.
* Updated `MLPBenchmarkTest` `MatmulParams` to be more comprehensive for
testing purposes.

## Fix for Persistent Circular Buffering

`getCircularBufferInsertionPosition` is relative to the `ComputeWarp`
for-loop. `compute_warp_insertion_position_` should be relative to the
number of for-loops in the stack.

* For persistent, the `ComputeWarp` is always the outer-most for-loop.
The stack is 1, and the offset is 0.
* For data-parallel, the ComputeWarp` is the third for-loop. The stack
is 3, and the offset is 2.

## WAR Sync Inserter Fix for Epilogue Loop

The `wgmma.commit` must always be inserted at the end of epilogue for
loop. The epilogue loop does not need `wgmma.wait` because it does not
have async inputs to guard. The `RAWSyncInserter` adds the `wgmma.wait`
before the output of wgmma expr is used by any other exprs.
Summary: Always commit after issuing a group of async exprs. Insert wait
only if inputs of async expr occur in the for-loop.