Enable cluster_dims in Hopper default matmul scheduler (#4317)
This sets `cluster_dims` to `{2, 1, 1}` when the X dimension of the grid
is divisible by 2. It seems this can give about a 1% speedup in such
cases.

## Possible future approach:

Note that another way to go about this would be the following: given
cluster dims `{cm, cn, 1}` first do a `cm` by `cn` tile split, _then_
split the CTA tile from the resulting inner dims. I.e. for problem size
`M, N, K` we would do
```
[M, N, K] -> [M / cm, N / cn, cm, cn, K]
  -> [M/cm, N/cn, cm/cta_m, cn/cta_n, K/cta_k, cta_m, cta_n, cta_k]
```
From there we can continue to split the warp and instruction tiles from
the cta tile. The advantage here is that we have used a nondivisible cga
tile split, so we can now merge and launch a different grid:
```
 -> [((M/cm) * (N/cn))/ncgas, BIDy(ncgas), BIDx((cm/cta_m * cn/cta_n)), ...]
```
This means we would launch a 2D grid for persistent of size `{cm/cta_m *
cn/cta_n, ncgas}`. BIDx now indicates the rank within the CGA and BIDy
is the CGA number, and we have an outer loop over "CGA tiles". This way
we can use CGAs whenever we want since the grid will always be a
multiple of the CGA size. cc @rdspring1