Fix resize scheduler with reshapes appearing after ref tensor (#4259)
Fix #4250

In the resize scheduler, the scheduling applied to the reference tensor
is propagated to all the other tensors by a backward traversal from the
reference loop domain over the exact graph. This means that the loop
domains of the other tensors need to be reachable from the reference.

The requirement does not hold when, for example, there's a reshape after
the reference. See the added test for a concrete example. More
generally, if the logical domain of a tensor is not reachable by just
traversing backward from the reference logical domain, there must be
something that should prevent the transform propagation. Note that since
the resize scheduler does not allow no additional `Resize` ops between
the ref and outputs, these additional ops between the ref and other
tensors must come from reshapes and should be "cancelable", i.e.,
setting the loop domain as the producer of the reshaped logical domain.
`prepareForBackwardTransformPropagation`, which is called just before
propagating ref transformations, searches for those tensors that have
additional transformations and make their loop domains look like the
reference logical domain.