Cleaning up scheduling of static repeat (#4325)
The resize scheduler automatically detects a sequence of ops to repeat a
tensor at a certain ID and slightly modifies the scheduling to reduce
redundant computations. This PR makes the analysis a little more
flexible so that it also works with a pattern appearing in a Llama
forward RoPE module.

Specifically, previously, the specific scheduling is only applied when a
sequence of `BroadcastOp, ExpandOp and ViewOp` are detected in this
order, just because that's how a repetition of a tensor is commonly
represented. However, the only op that is absolutely necessary is the
final reshape. As long as it meets the patterns for repetition, it
should be sufficient to apply the scheduling. In fact, in a segment of a
Llama RoPE forward, there's a segment input that has a broadcast ID,
which is then expanded inside the segment and merged to realize a
repetition. This case is not detectable as the segment lacks a
BroadcastOp in the current main but is detected with this PR.