Limit grid traversal factor to prevent suboptimal CTA order
I noticed a case where we had a 13x10 tile grid. We rightly chose
row-major `cta_order` since 10 is smaller than 13 so it should be the
fast dimension. However we then set `grid_traversal_factor` to 13. When
we set the `grid_traversal_factor` equal to the size of the other
dimension, the traversal order is no different than swapping to column
major, which is no longer optimal. Instead, we should only consider
swizzles that are smaller than the number of tiles in the slow
dimension.

Using `benchmarks/python/test_matmul.py`:
Before: 61.8% of cublas
After: 67.4% of cublas

This is a 9% speedup.