Allocate and Deallocate HostIr insertion in FusionKernelRuntime (#4329)
Follow-up PR to https://github.com/NVIDIA/Fuser/pull/4286. The PR will
insert allocate ops for every LaunchKernel output, and also insert a
Deallocate right after the last use of every input expr in the Hostir
container. It adds a test to check the number of Deallocate ops and the
max memory usage is correct for an example fusion as well.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>