Ignore device parallelization on reduction axis in sharding propagation. (#4353)
When propagating shardings to a tensorview, we skip device parallel
types already present on that tensorview. We ignore reduction dimensions
for this.

I discovered a bug in the backpropagation of shardings. Since I was
passing an empty set as selected parallel types to
`selectiveReorderDIDToFront`, no backpropagation was occuring (An
earlier version of propagation used excluded parallel types instead of
selected parallel types which led to this bug). This got masked in the
`ShardingTest.ResidualAdd` since the allocation domain of tvs was not
set and `getShardedLogicalAxis` uses that.