Don't link lib_kernels.so against codegen_internal (#4378)
That file is the result of tests/cpp/rng_kernels.cu, which does not use
anything from nvfuser so it does not need to be linked against nvfuser.
Doing so causes a multiply defined symbol, leading to this type of asan
error:

https://github.com/google/sanitizers/wiki/AddressSanitizerOneDefinitionRuleViolation.

NOTE: It seems `$NVFUSER_TESTS` is never set in CMakeLists.txt. I am
unsure what the original intended name was for this file but I think we
should just hardcode that name instead of the current version which
leads to `lib_kernels.so` being generated..

Fixes #4369.