Fix reference reordering in resize scheduler (#4356)
In the resize scheduler, there's an assumption that, when reordering the
reference tensor using an input tensor, the innermost ID of the
reference is reachable from the IDs of the input tensor. That's actually
not the case when the innermost ID is repeated, which hits [this
assertion](https://github.com/NVIDIA/Fuser/blob/main/csrc/scheduler/resize.cpp#L509).

This PR ensures the innermost reference ID always stays at the innermost
position by excluding it from reordering.