Make sure only active ID exprs are considered (#4442)
Currently, IterDomain::definition() is not guaranteed to return an
active expr with respect to the associated TensorDomain, i.e., may not
be included in TensorDomain::allExprs. That happens when
TensorDomain::setLoopDomain cuts off IDs that are not in the paths
between the logical/root/allocation/loop domains. We may want to remove
dangling definitions, but for now, this PR fixes an error by explicitly
checking if an ID is indeed an active ID.

Fixes #4444 

Issue for dangling exprs: #4445