New non-divisible split analysis for non-straightline logical-to-loop transformations (#4390)
We are now allowing to schedule loop IDs not just as a (indirect)
consumers of logical IDs but they can be producers as well. For example,
that's how we reduce redundant computations when a fusion ends with a
repeat op:

```
// t0: [i0, i1]
t1 = repeat(t0, {2, 1}); // Repeat the outer dimension twice
// t1: [i0*2, i1]
```

In this case, `t1` is represented with a reshape:

```
root: i0, i2(2), i1
logical:  i0*i2(2), i1
```

In the resize scheduler, when an ending repeat is detected, we first
cancel the reshape in the loop domain:

```
loop: [i0, i2(2), i1]
```

Then move `i2` to the outermost:

```
loop: [i2(2), i0, i1]
```

Then we schedule the inner two IDs of `t1` and the rest of the tensors
in the fusion together. See #3716 for more details.

When a tensor is scheduled this way, the existing analysis of
non-divisible splits is not adequate as it only always looks at ID exprs
from the loop IDs to backwards. In the case of the above example, the
predicate indexing path from loop to logical would end with the merge of
`i0` and `i2`, however, that merge would not be considered. Also, it
implicitly assumes that the IDs with no defining expression are always
predicated in order to minimize extra predicates. That assumption made
sense when the original analysis was implemented but no longer the case.

This PR reimplements the necessary analysis for `TensorIndexer` and does
not assume the straight-line logical-to-loop dependency.

One concern is that the same indexing traversal is done twice for this
non-divisible analysis and the actual predicate generation
(`TensorIndexer::getPredicates`). I didn't dive into too much, but
because of the inter dependencies between the non-divisible analysis,
predicate elimination, and TensorIndexer, I didn't find an easy way to
avoid that. Note that it's also the case with the current
`NonDivisibleSplitInfo`, but it's just an `IterVisitor` traversal,
whereas for IdModel it's a BFS traversal.

Fixes #4376