Cleanup matmul scheduler (#4457)
Today, we have a `MatmulScheduler` which is the scheduler entry for the
registry, and `MultipleMatmulScheduler` which is a way for us to
organize our code that applies the schedule to the fusion
(`MatmulScheduler::schedule`). These two codes are relatively
independent, so I still intend to keep this structure. The things I
changed are:

- What was `MultipleMatmulScheduler`,
`AmpereMinusMultipleMatmulScheduler`,
`HopperPlusMultipleMatmulScheduler` now goes to a new namespace
`schedule_matmul`, with a shorter name `Common`, `AmpereMinus`, and
`HopperPlus`.
- `matmul.{h,cpp}` and `multi_matmul.{h,cpp}` are merged together.
- `multi_matmul_ampere-.{h,cpp}` are renamed as
`matmul_ampere-.{h,cpp}`.
- `multi_matmul_hopper+.{h,cpp}` are renamed as
`matmul_hopper+.{h,cpp}`.