Reject cyclic reshape (#4437)
WAR for #4415 

It is generally assumed that the ID graph is acyclic. We do support
cycles due to resize in a limited form. Cyclic reshapes can be a
problem, although I'm not sure where all the places we implicitly
assumed the property.
[propagateReshapeTransforms](https://github.com/NVIDIA/Fuser/blob/main/csrc/scheduler/utils.cpp#L2250)
is one of them.

This PR rejects a segment if the pattern is detected.

Closes #4415