Fix preferred merge (#4469)
Fixes #4459 

The bug didn't result in any error and in fact allowed fusing a concat
op with its preceding pad ops, which improved some of the RoPE
benchmarks. With this fix, it appears that the segmenter fails to fuse
them, so this results in a performance loss. I'll have a follow-up
[PR](https://github.com/NVIDIA/Fuser/pull/4481) to recover it.