Add missing non-divisible predicates in TensorIndexer (#4236)
Fixes #4218 

Because we use the almost-exact graph for indexing traversal, some
non-divisible splits that should be predicated may not be included in
the traversal, thus missing those predicates in the final code. That's
the reason of the validation failure as discussed in #4218.

This PR addresses the problem by grabbing all such potentially missing
splits and ensure they are properly predicated. One complication is that
since those splits do not appear in the original indexing traversal, the
IDs to predicate may not have corresponding indices in the index map. In
this PR, I simply do another BFS traversal to fill the missing mappings.

Ideally, we should only do the traversal once, but until the first
traversal is done, we don't know which splits are missing. So, I think
as long as the almost-exact graph is used, second traversal cannot be
avoided.

Alternatively, we might be able to use the exact graph instead of the
almost-exact graph. With the exact graph, all necessary non-divisible
splits should be traversed, so this problem should not exist. Changing
the graph type just for this problem doesn't seem like a good idea to
me, so I didn't take this direction.

### Code diff
I enabled TensorIndexer for `TMASimpleLdstTest`. Locally ran the code
diff checker, while many tests resulted in showing some diffs, I believe
all of them are basically variations of the change as shown below:

```
1: TMASimpleLdstTest.Load/3D_NoSwizzle___half
  Kernel 1 CUDA PTX    53997da5d Diff 8b772f39b -1 +1index type: int registers: 15 gmem: 3 static smem: 0 stack frame: 8 spill stores: 0 spill loads: 0
--- 53997da5d

+++ 8b772f39b

@@ -18,11 +18,11 @@

   nvfuser_index_t i8;
   i8 = 5 * i6;
   Array<int, 3, 1> a9;
   a9 = Array<int, 3, 1>{(int32_t)(i7), (int32_t)(i8), (int32_t)(i5)};
   nvfuser_index_t i10;
-  i10 = (((nvfuser_index_t)threadIdx.x) / 8) % 5;
+  i10 = ((nvfuser_index_t)threadIdx.x) / 8;
   nvfuser_index_t i11;
   i11 = ((nvfuser_index_t)threadIdx.x) % 8;
   __half* T1 = reinterpret_cast<__half*>(array + smem_offset + 0);
   uint64_t* T3 = reinterpret_cast<uint64_t*>(array + smem_offset + 80);
   mbarrier::init(toSmem(T3), 1U);
```

So, TensorIndexer doesn't have `% 5`. I confirmed that TensorIndexer
generates correct indices, so I believe these changes are benign.

`test_issue1872` is also flagged as follows:

```
tests/python/test_python_frontend.py::TestNvFuserFrontend::test_issue1872
  Kernel 1 CUDA PTX    0dc45a49 Diff 4c1d98ca -1 +3index type: int registers: 8 gmem: 3 static smem: 0 stack frame: 0 spill stores: 0 spill loads: 0
--- 0dc45a49

+++ 4c1d98ca

@@ -8,7 +8,9 @@

     T1[0]
        = T0[0];
     T5[i0]
        = __float2half(T1[0]);
   }
-  loadLocalToGlobal<__half, /*vec_size=*/2, /*is_volatile=*/false>( &T2[0], &T5[0]);
+  if ((((nvfuser_index_t)threadIdx.x) < 1)) {
+    loadLocalToGlobal<__half, /*vec_size=*/2, /*is_volatile=*/false>( &T2[0], &T5[0]);
+  }
 }
```

The current main does not generate a predicate, which is correct but
redundant. All threads just do the same computation and stores the same
result to the same location. The change looks expected.