Make sure concat is fused with pads of inputs (#4481)
The bug fixed in #4469 resulted in segmenting a concat op with its input
pad ops. This can result in passing padded tensors between a preceding
segment and the segment for the concat op. I don't see any reason that
we may want to have these ops in different segments as they are
originally from the same user-level op (`cat`). This PR tries to avoid
the segmentation by merging a concat with input pad groups.

I confirmed that after this PR, the resulting generated code for the
RoPE benchmark is unchanged compared to the main branch prior to #4469.