Merge batch dims before grid parallelization in Hopper matmul (#4464)
Fixes an issue that appeared in CI while testing #4458. In that case, we
limited the CTA K dimension to smaller than the problem K which caused
the number of K loop iterations to go from 1 to 2 and caused circular
buffering to be enabled in a BMM example. This exposed a problem where
the batch dimension was picked up as the circular buffer dimension. This
PR removes that outer batch dimension loop.

This merges the batch dim(s) as the outer dim outside of BIDy for
non-persistent. For persistent matmul, we merge the batch dim as outer
before doing the persistent split. This prevents us having another outer
loop which interfered with circular buffering analysis.

TODO: add a test that exercises circular buffering with bmm in both
persistent and non-persistent use cases.