Skip short-circuit for trivial loops, use ID producer extent (#4506)
I found this when enabling persistent CTAs in Hopper matmul by default
in #4498. This test has static sizes and there are at most 8 tiles due
to the problem size:

https://github.com/NVIDIA/Fuser/blob/6b5de266ab25fbb61930f1fc935a04764ce19aa9/tests/cpp/test_matmul_scheduler.cpp#L1037-L1050

This means we generate a persistent loop with only a single iteration.
Since we have static shapes the extent of the persistent loop is a
constant 1. In our lowering parlance, this is a "trivial loop". Such
loops are not actually generated, so we should not create an ITE where
the loop index variable appears in the predicate. This PR does that, and
also strengthens the check by directly checking the ID expressions
instead of the definition of the loop's `stop` scalar.