Enable persistent CTAs by default in Hopper matmul heuristic (#4498)
Currently we limit the number of circular buffering stages in our
heuristic to the number of stages per tile. Since mbarriers have an smem
overhead as well as a latency overhead due to
initialization/invalidation we should not use more stages than needed
for a given problem. However, when we use persistent CTAs we are able to
circular buffer across _tiles_ as well, so the limit should be increased
to the number of stages per tile times the max number of tiles per SM.
That is what this PR does. This gives a rather large boost to perf in
cases with small K with the most extreme cases being K <= 16 since it
enables circular buffering in those cases where none is possible for
non-persistent kernels.

This PR also includes a minor refactor where I created
`maximizeHopperOperandStages` to try and maximize the number of stages
used given the CTA tile size and available smem. This is be used before
`limitCircularBufferingSmemOperands` to automatically set the number of
stages for both the default heuristic and to post-process plugin
configs.

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>
Co-authored-by: Jingyue Wu <wujingyue@gmail.com>
Co-authored-by: samnordmann <snordmann@nvidia.com>