Map CGA dims using 3D grid in persistent Hopper matmul
Stacked on #4405.

Given something like `cluster_dims = {4, 2, 1}`, this PR creates a 3D
grid of size `{4, 2, 66}` (assuming 132 SMs). This makes it trivial to
launch the proper number of CTAs and can simplify indexing for finding
the position within the CGA. That is not currently needed but it would
be if we want to do explicit multicast or broadcasting within a CGA.