consider all for-loops between circular buffer loop and slice stage position (#4524)
**Issue:** Predicate for `MBarrierWaitParity` is missing when multiple
computation warp groups are used. For example, the following
`MBarrierWaitParity` should have a predicate using `i101` and
`threadIdx.y`
```
    FOR i101 in iS24{13}:
      FOR threadIdx.y in ithreadIdx.y25{2}:
        T2_l_float[iblockIdx.x17{148}, iS18{13}, ithreadIdx.y19{2}, iUR15{4}, ithreadIdx.x5{128}] ca_pos( 3 ) produce_pos( 2 ) = ALLOCATE(buffer=T2_l_float[iblockIdx.x17{148}, iS18{13}, ithreadIdx.y19{2}, iUR15{4}, ithreadIdx.x5{128}] ca_pos( 3 ) produce_pos( 2 ), mem_type=register, size=4, zero_init=false, resets_to_zero=false)
        MBarrierWaitParity(T4_s[( ( threadIdx.y + ( 2 * i101 ) ) % 6 )] view( T4 ), i1277)
```
**Why:**
(1) Previous impementation only checks the last for-loop to identify the
location of `MBarrierWaitParity`. However, when ping-pong (or more
generally `slice_stage_position`) is used, extra for-loops are added,
e.g. `FOR threadIdx.y in ithreadIdx.y25{2}` in this case.
(2) It was not triggered in previous test becuase the only test using
multiple computation warp groups was designed before adding 1d tma
predicate, all splits in that test are divisible.

**Fix:**
(1) Check all the stacked `for loops`.
(1) Added check to ensure there is no unpaired `PredicateType::OneDimTmaLoadExpectArrive`